% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internals.R
\name{check_for_quotes}
\alias{check_for_quotes}
\title{Regexp for quotations marks around the text (\code{x})}
\usage{
check_for_quotes(x)
}
\arguments{
\item{x}{Text to check}
}
\value{
\code{logical}, indicator whether \code{x} is within quotation marks
}
\description{
Checks whether \code{x} is within quotation marks (either \' or \").
Internal function, not exported.
}
\examples{
check_for_quotes('\\'C:\\\\Users\\'')
check_for_quotes('C:\\\\Users')
}
