% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{path_parse}
\alias{path_parse}
\title{Parse selected "Windows" path to an "R-usable" one}
\usage{
path_parse()
}
\description{
Parse selected "Windows" path to an "R-usable" one
}
\examples{
# Not run: test on the following texts:
C:\\Users\\me
'C:\\Users\\me'
"C:\\Users\\me"
C:\\\\Users\\\\me
C:/Users/me
}
